TASK: Allow to disable thumbnails for news lists
The fetching if a thumbnail from the news is an expensive operation. If you
do not need thumbnails in the list view (either with the news list or in a
news/category folder), you can now switch off thumbnails.

This speeds up rendering up to 10 times by avoiding the costly FlowQuery
used to look for (potential) thumbnails.