BUGFIX: Fix wrong results of merge
This fixes the merge of the bellicon/3.1 branch into master, which
I did with too little care…

It moves two files from the TypoScript folder to the Fusion folder
and corrects the collection of news items to use the new collector
object again.

In addition it renames .ts2 files to use the .fusion extension.